requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
python-dotenv
psycopg2-binary
dj-database-url
Pillow
django-storages
boto3
